test(refactor): Deleting unnecessary test and relocating test to unittest.
