+++ test/streams/test_audio.py
@@ -143,7 +143,7 @@ def test_node_save_audio_stream(audio_node_stream):
     audio_node_stream.start()
 
     # Wait to generate files
-    time.sleep(10)
+    time.sleep(3)
 
     audio_node_stream.shutdown()
     audio_node_stream.join()
