+++ test/streams/test_transfer.py
@@ -30,11 +30,11 @@ pytestmark = [pytest.mark.slow, pytest.mark.timeout(600)]
 
 NAME_CLASS_MAP = {
     "vn": VideoNode,
-    "img_n": ImageNode,
-    "tn": TabularNode,
-    "an": AudioNode,
+    # "img_n": ImageNode,
+    # "tn": TabularNode,
+    # "an": AudioNode,
 }
-NUM_OF_WORKERS = 3
+NUM_OF_WORKERS = 2
 
 
 @pytest.fixture
@@ -180,18 +180,18 @@ def test_worker_data_archiving(worker):
 @pytest.mark.parametrize(
     "config_manager, expected_number_of_folders",
     [
-        (lazy_fixture("single_worker_manager"), 1),
-        (lazy_fixture("multiple_worker_manager"), NUM_OF_WORKERS),
+        # (lazy_fixture("single_worker_manager"), 1),
+        # (lazy_fixture("multiple_worker_manager"), NUM_OF_WORKERS),
         pytest.param(
             lazy_fixture("dockered_single_worker_manager"),
             1,
             marks=linux_run_only,
         ),
-        pytest.param(
-            lazy_fixture("dockered_multiple_worker_manager"),
-            NUM_OF_WORKERS,
-            marks=linux_run_only,
-        ),
+        # pytest.param(
+        #     lazy_fixture("dockered_multiple_worker_manager"),
+        #     NUM_OF_WORKERS,
+        #     marks=linux_run_only,
+        # ),
     ],
 )
 def test_manager_worker_data_transfer(config_manager, expected_number_of_folders):
