+++ test/test_manager.py
@@ -1,6 +1,12 @@
 from pathlib import Path
 
+import pytest
+from pytest_lazyfixture import lazy_fixture
+
+import chimerapy as cp
 from chimerapy.manager import Manager
+
+from .conftest import GenNode, ConsumeNode, linux_run_only
 
 
 def test_manager_logdir_string():
